+++ frb_example/pure_dart/rust/src/api/impl_trait.rs
@@ -1,3 +1,4 @@
+use crate::frb_generated::SimpleTraitForDynTwinNormalImpl;
 use flutter_rust_bridge::frb;
 
 pub trait SimpleTraitTwinNormal {
@@ -38,4 +39,36 @@ impl SimpleTraitTwinNormal for StructTwoWithTraitTwinNormal {
     fn simple_trait_fn_receiver_borrow_twin_normal(&self) -> i32 {
         self.two * 2
     }
+}
+
+#[frb(generate_impl_enum)]
+pub trait SimpleTraitForDynTwinNormal {
+    fn simple_method_twin_normal(&self) -> i32;
+}
+
+#[frb(opaque)]
+pub struct StructOneWithTraitForDynTwinNormal {
+    pub one: i32,
+}
+
+impl SimpleTraitForDynTwinNormal for StructOneWithTraitForDynTwinNormal {
+    fn simple_method_twin_normal(&self) -> i32 {
+        self.one
+    }
+}
+
+#[frb(opaque)]
+pub struct StructTwoWithTraitForDynTwinNormal {
+    pub two: i32,
+}
+
+impl SimpleTraitForDynTwinNormal for StructTwoWithTraitForDynTwinNormal {
+    fn simple_method_twin_normal(&self) -> i32 {
+        self.two * 2
+    }
+}
+
+pub fn func_arg_trait_impl_twin_normal(arg: SimpleTraitForDynTwinNormalImpl) -> i32 {
+    let arg = arg.blocking_read();
+    arg.simple_method_twin_normal()
 }
